refactor: route console I/O through System, drop Console trait
Console (renamed from Stdio) now delegates all stdio primitives
(print, prompt, raw mode) to System instead of calling std::io
directly, and the Console trait is gone since Stdio was its only
implementor. Tests build a real Console over SystemMock instead of
a separate hand-maintained ConsoleMock.
