+++ src/commands/list_image_command.rs
@@ -39,7 +39,7 @@ pub struct ListImageCommand {
 }
 
 impl Command for ListImageCommand {
-    fn run(&self, console: &mut dyn Console, context: &Context) -> Result<()> {
+    fn run(&self, console: &mut Console<'_>, context: &Context) -> Result<()> {
         let images = fetch_image_list(console, context.get_env());
 
         let mut view = TableView::new();
