+++ src/commands/prune_command.rs
@@ -17,7 +17,7 @@ pub struct PruneCommand {
 }
 
 impl Command for PruneCommand {
-    fn run(&self, console: &mut dyn Console, context: &commands::Context) -> Result<()> {
+    fn run(&self, console: &mut Console<'_>, context: &commands::Context) -> Result<()> {
         let env = context.get_env();
 
         // Calculate size
