+++ src/commands/run_command.rs
@@ -38,7 +38,7 @@ pub struct RunCommand {
 }
 
 impl Command for RunCommand {
-    fn run(&self, console: &mut dyn Console, context: &commands::Context) -> Result<()> {
+    fn run(&self, console: &mut Console<'_>, context: &commands::Context) -> Result<()> {
         self.create_cmd.run(console, context)?;
         commands::SshCommand {
             target: Target::from_instance_name(self.create_cmd.instance_name.value.clone()),
