+++ src/commands/stop_command.rs
@@ -41,7 +41,7 @@ pub struct StopCommand {
 }
 
 impl Command for StopCommand {
-    fn run(&self, console: &mut dyn Console, context: &commands::Context) -> Result<()> {
+    fn run(&self, console: &mut Console<'_>, context: &commands::Context) -> Result<()> {
         let instance_store = context.get_instance_store();
 
         if !self.all {
@@ -84,7 +84,6 @@ mod tests {
     use crate::instance::InstanceStoreMock;
     use crate::models::Environment;
     use crate::platform::SystemMock;
-    use crate::view::ConsoleMock;
     use std::rc::Rc;
 
     #[test]
@@ -94,7 +93,8 @@ mod tests {
 
     #[test]
     fn test_reject_empty_instance_list_without_all() {
-        let console = &mut ConsoleMock::new();
+        let system = SystemMock::new();
+        let console = &mut Console::new(&system);
         let env = Environment::new(
             "myuser".to_string(),
             String::new(),
@@ -122,7 +122,8 @@ mod tests {
 
     #[test]
     fn test_allow_empty_instance_list_with_all() {
-        let console = &mut ConsoleMock::new();
+        let system = SystemMock::new();
+        let console = &mut Console::new(&system);
         let env = Environment::new(
             "myuser".to_string(),
             String::new(),
