+++ src/platform.rs
@@ -1,6 +1,8 @@
+mod stream;
 mod system;
 mod system_mock;
 
+pub use stream::*;
 pub use system::*;
 #[cfg(test)]
 pub use system_mock::tests::SystemMock;
