+++ src/platform/stream.rs
@@ -1,0 +1,5 @@
+#[derive(Clone, Copy)]
+pub enum Stream {
+    Stdout,
+    Stderr,
+}
