+++ src/view.rs
@@ -2,11 +2,9 @@ mod animation;
 mod async_transfer_view;
 mod confirm_dialog;
 mod console;
-mod console_mock;
 mod map_view;
 mod progress_bar;
 mod spinner;
-mod stdio;
 mod table_view;
 mod time_duration;
 mod transfer_view;
@@ -15,12 +13,9 @@ pub use animation::*;
 pub use async_transfer_view::*;
 pub use confirm_dialog::*;
 pub use console::*;
-#[cfg(test)]
-pub use console_mock::tests::ConsoleMock;
 pub use map_view::*;
 pub use progress_bar::*;
 pub use spinner::*;
-pub use stdio::*;
 pub use table_view::*;
 pub use time_duration::*;
 pub use transfer_view::*;
