+++ src/view/confirm_dialog.rs
@@ -11,7 +11,7 @@ impl ConfirmDialog {
         }
     }
 
-    pub fn confirm(&self, console: &mut dyn Console) -> bool {
+    pub fn confirm(&self, console: &mut Console<'_>) -> bool {
         let reply = console.prompt(&format!("{} [y/N]: ", self.message));
         matches!(reply.to_lowercase().as_str(), "y" | "yes")
     }
@@ -20,33 +20,37 @@ impl ConfirmDialog {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::view::ConsoleMock;
+    use crate::platform::SystemMock;
 
     #[test]
     fn test_confirm_accepts_y() {
-        let console = &mut ConsoleMock::new();
-        console.push_input("y");
+        let system = SystemMock::new();
+        system.push_input("y");
+        let console = &mut Console::new(&system);
         assert!(ConfirmDialog::new("Proceed?").confirm(console));
     }
 
     #[test]
     fn test_confirm_accepts_yes_case_insensitive() {
-        let console = &mut ConsoleMock::new();
-        console.push_input("Yes");
+        let system = SystemMock::new();
+        system.push_input("Yes");
+        let console = &mut Console::new(&system);
         assert!(ConfirmDialog::new("Proceed?").confirm(console));
     }
 
     #[test]
     fn test_confirm_rejects_empty_reply() {
-        let console = &mut ConsoleMock::new();
-        console.push_input("");
+        let system = SystemMock::new();
+        system.push_input("");
+        let console = &mut Console::new(&system);
         assert!(!ConfirmDialog::new("Proceed?").confirm(console));
     }
 
     #[test]
     fn test_confirm_rejects_anything_else() {
-        let console = &mut ConsoleMock::new();
-        console.push_input("n");
+        let system = SystemMock::new();
+        system.push_input("n");
+        let console = &mut Console::new(&system);
         assert!(!ConfirmDialog::new("Proceed?").confirm(console));
     }
 }
